Updated index.html
update the code to improve the import functionality and enhance the settings UI, particularly for the "Import Data" section. I'll modify the settings popup to show the chosen file name and improve the overall styling of the settings section. Since this is an update to a previous artifact, I'll use the same artifact_id from the conversation history, but I don't see a prior artifact_id in this conversation,

Here are the changes:

Add a display for the chosen file name below the Import Data input
Style improvements for the settings popup to make it more organized
Update the importData function to show the chosen file name
Add some spacing and visual hierarchy to the settings form